[Change] move classes to correct packages
